+++ NEWS.md
@@ -4,6 +4,7 @@ ClimaAnalysis.jl Release Notes
 v0.5.6
 ------
 - Fix finding variables with name like `clwup_1m_40s_inst.nc` (composed period).
+- Add support for weighted averages in `average_lat`.
 
 v0.5.5
 ------